+++ Code/Dia03/Lectura/Figura.java
@@ -3,7 +3,7 @@ package Dia03.Lectura;
 public abstract class Figura {
     private String name;
     
-    public abstract int getPerimetro();
+    public abstract double getPerimetro();
 
     protected void setName(String name){
         this.name = name;
